pyproject.toml
[tool.poetry]
name = "hausportal"
version = "0.1.0"
description = "Haus- und Mieterportal Backend"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = "^0.128.0"
uvicorn = {extras = ["standard"], version = "^0.40.0"}
sqlalchemy = "^2.0"
psycopg = {extras = ["binary"], version = "^3.2"}
alembic = "^1.18"
pydantic-settings = "^2.12"
python-multipart = "^0.0.21"
pyjwt = "^2.10"
passlib = {extras = ["bcrypt"], version = "^1.7"}
httpx = "^0.28"
bcrypt = "<4"
email-validator = "^2.3.0"
boto3 = "^1.42.30"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0"
ruff = "^0.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
